requirements.txt
#Propósito: Lista las dependencias de Python necesarias para el proyecto

pyautogui
pillow
tkcap
pydicom
img2pdf
opencv_python
matplotlib
pandas
tensorflow==2.16.2
python-xlib
pytest
